requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
discord.py==2.1.0
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
pytz==2022.7.1
requests==2.28.2
urllib3==1.26.14
yarg==0.1.9
yarl==1.8.2
